// src/models/comunes/LocalStorageData.ts
export interface CommerceData {
    rut: string;
    commerce_rut: string;
    email: string;
    // commerce_code: number;
    // commerce_id: number;
    // contract_id: string;
    first_name: string;
    last_name: string;
    // second_last_name: string;
    // business_name: string;
    // taxpayer_type: string;
    fantasy_name: string;
    // area_of_work: string;
    // line_of_business: string;
    // address: string;
    // address_type: string;
    // address_number: string;
    // town: string;
    // phone: string;
    // cellphone: string;
    // birthdate: string;
    // region_code: string;
    // city_code: string;
    // district_code: string;
    // last_modified: string;
    // creation_date: string;
    // contract_type: string;
    // logic_balance_type: string;
    // collector: string;
    // district: string;
    // region: string;
    // city: string;
    // reinstatement_date: string;
    // latitude: string;
    // longitude: string;
    // deposit_email: string;
    // rut_bill_branch_technology: string;
  }
  